requirements.txt
asgiref==3.7.1
attrs==23.1.0
black==23.3.0
click==8.1.3
colorama==0.4.6
Django==4.2.1
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
exceptiongroup==1.1.1
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.0.0
pyrsistent==0.19.3
pytest==7.3.1
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.1
tzdata==2023.3
uritemplate==4.1.1
